refactor(config): extrair conversores de settings.py para _coercions.py
Extrai as 14 funções `_to_*` (exceto `_to_log_level`), o alias
`_to_optional_pos_int` e as duas regexes `_MODEL_SLUG_RE` /
`_ACTIVITY_DEPLOYMENT_RE` de `deile/config/settings.py` para o módulo
irmão `deile/config/_coercions.py`.

`settings.py` mantém um shim de re-export (`from ._coercions import ...`)
que preserva a superfície pública integralmente — os 40+ importadores
externos e os 9 testes que importam conversores diretamente de
`deile.config.settings` funcionam sem alteração.

`_coercions.py` é módulo-folha: sem imports top-level de deile. As duas
dependências internas (`is_valid_dispatcher`, `is_valid_effort`) permanecem
como imports lazy dentro das funções, mantendo o grafo acíclico.

Ficam em `settings.py`: `LogLevel` + `_to_log_level` (evita ciclo via
enum), `_normalize_path_for_comparison` e `_is_project_layer_trusted`
(chamados por nome no corpo do módulo / importados por teste).

LOC: `_coercions.py` = 330 (< ~340), `settings.py` = 1444 (< ~1500).
Singleton `get_settings() is get_settings()` verificado. Grafo acíclico
confirmado. 203 testes do subconjunto impactado passam.

Refs #751
